requirements.txt
web3
eth-account
eth-typing
python-dotenv
eth_abi
click
pytest
setuptools
pynvml
psutil